build in check for tweet recurrence flag
The Library Team requested that we be able to have some tweets allowed
to recur. Therefore, we added a column called "Recurrence" to the Tweet
Entry form, which takes "Yes" if the tweet is allowed to recur and "No"
if not. The new check occurs within `post_tweet_library()` and removes tweets with `Recurrence == "Yes"` from the
list IDs to be omitted due to having been posted once already.